requirements.txt
boto3
cfn_flip
flask
jsondiff
pyyaml
